Information correction (#28582)
Eliminated the line  (<code>rotate()</code> works the same way that <code>skewX()</code> and <code>skewY()</code> do).
It is incorrect. SkewX and SkewY do not work the same as rotate, perhaps in the sense that they take an argument ranging from 0 to 360deg, but they do not have the same effect.